relocate util funcs from CodeGen to CodeGen.Util
