+++ package.yaml
@@ -16,6 +16,7 @@ library:
     - Hibiscus.Parser
     - Hibiscus.CodeGen.Type.DataType
     - Hibiscus.CodeGen.Type
+    - Hibiscus.CodeGen.Util
     - Hibiscus.CodeGen
     - Hibiscus.Asm
     - Hibiscus.Type4plus
